Simplify the bus network structure
Replace the buses dictionary, where the bus number was used
as the key, with an array. The array contains dictionaries.
Each dictionary contains the name of the bus, which consists
of both the number and the route, and the direct and reverse
routes. It is easier to display the buses in an ordered fashion
on the client side in this way.