Simplify the application - remove the favorites
Start the simplification of the application. That is, remove
favorites functionality added before and drop the implementation
of the search feature. This way, the app should be simpler to
maintain. But also, the main reason is that it's difficult (for
me, at least) to come up with a good solution to the features
I've had in mind.
The target is to have a very simple application, where the user
can select the bus, then the bus station and then see the
timetable for that station without any extra functionality.